8.23

+++ Interface_Teacher_Testmanage.aspx.cs
@@ -101,7 +101,7 @@ public partial class Interface_Teache_Testmanage : System.Web.UI.Page
 
             string xiangdui = @"~\test\";//设置文件相对网站根目录的保存路径 ，~号表示当前目录，在此表示根目录下的images文件夹
             string juedui = Server.MapPath("~\\test\\");//设置文件保存的本地目录绝对路径，对于路径中的字符“＼”在字符串中必须以“＼＼”表示，因为“＼”为特殊字符。或者可以使用上一行的给路径前面加上＠
-            string newFileName =juedui+strName;
+            string newFileName = juedui + strName;
             if (FileUpload1.HasFile)//验证 FileUpload 控件确实包含文件
             {
                 String[] allowedExtensions = { ".gif", ".png", ".bmp", ".docx", ".txt" };
@@ -115,25 +115,25 @@ public partial class Interface_Teache_Testmanage : System.Web.UI.Page
             }
             if (fileOK)
             {
-                    // 判定该路径是否存在
-                    if (!Directory.Exists(juedui))
-                        Directory.CreateDirectory(juedui);
-                    // Label3.Text = newFileName;     //为了能看清楚我们提取出来的图片地址，在这使用label
-                    /* Label4.Text = "<b>原文件路径：</b>" + FileUpload1.PostedFile.FileName + "<br />" +
-                                       "<b>文件大小：</b>" + FileUpload1.PostedFile.ContentLength + "字节<br />" +
-                                       "<b>文件类型：</b>" + FileUpload1.PostedFile.ContentType + "<br />";
-                                       */
-                    // Label3.Text = xiangdui + newName + kzm;
-                    // Label6.Text = "文件上传成功.";
-                    //Label3.Text = strName;
+                // 判定该路径是否存在
+                if (!Directory.Exists(juedui))
+                    Directory.CreateDirectory(juedui);
+                // Label3.Text = newFileName;     //为了能看清楚我们提取出来的图片地址，在这使用label
+                /* Label4.Text = "<b>原文件路径：</b>" + FileUpload1.PostedFile.FileName + "<br />" +
+                                   "<b>文件大小：</b>" + FileUpload1.PostedFile.ContentLength + "字节<br />" +
+                                   "<b>文件类型：</b>" + FileUpload1.PostedFile.ContentType + "<br />";
+                                   */
+                // Label3.Text = xiangdui + newName + kzm;
+                // Label6.Text = "文件上传成功.";
+                //Label3.Text = strName;
 
-                   Response.Write("<script>alert('文件上传成功')</script>");
-                    FileUpload1.PostedFile.SaveAs(newFileName);//将图片存储到服务器上
+                Response.Write("<script>alert('文件上传成功')</script>");
+                FileUpload1.PostedFile.SaveAs(newFileName);//将图片存储到服务器上
             }
             else
             {
                 Response.Write("<script>alert('不能上传该类型文件')</script>");
-               // Label6.Text = "只能够上传图片文件.";
+                // Label6.Text = "只能够上传图片文件.";
             }
         }
     }
@@ -176,6 +176,24 @@ public partial class Interface_Teache_Testmanage : System.Web.UI.Page
             if ((String)Session["SelectedFile"] != "")
             {
                 string FullFileName = Session["SelectedFile"].ToString();
+                DownLoadFile(FullFileName);
+            }
+        }
+        else
+        {
+            Response.Write("<script>alert('请先选择要下载的文件')</script>");
+        }
+    }
+
+    protected void Button8_Click(object sender, EventArgs e)
+    {
+
+        // 判断是否选择了文件名
+        if (ListBox1.SelectedValue != "")
+        {
+            if ((String)Session["SelectedFile"] != "")
+            {
+                string FullFileName = Session["SelectedFile"].ToString();
                 DeleteFile(FullFileName);
                 Response.Redirect(Request.Url.PathAndQuery.ToString());
             }
@@ -186,34 +204,8 @@ public partial class Interface_Teache_Testmanage : System.Web.UI.Page
         }
     }
 
-    protected void Button7_Click1(object sender, EventArgs e)
-    {
-        // 判断是否选择了文件名
-        if (ListBox1.SelectedValue != "")
-        {
-            if ((String)Session["SelectedFile"] != "")
-            {
-                string FullFileName = Session["SelectedFile"].ToString();
-                DownLoadFile(FullFileName);
-            }
-        }
-        else
-        {
-            Response.Write("<script>alert('请先选择要下载的文件')</script>");
-        }
-    }
-
-    protected void ListBox1_SelectedIndexChanged1(object sender, EventArgs e)
+    protected void ListBox1_SelectedIndexChanged(object sender, EventArgs e)
     {
         Session["SelectedFile"] = ListBox1.SelectedValue.ToString();
-    }
-
-
-
-
-
-    protected void Button11_Click(object sender, EventArgs e)
-    {
-
     }
 }
